Remove redux-thunk by merging validation and common reducer Rename common reducer to core
