+++ packages/core/package.json
@@ -39,7 +39,6 @@
     "react": "^16.2.0",
     "react-redux": "^5.0.6",
     "redux": "^3.7.2",
-    "redux-thunk": "^2.2.0",
     "uuid": "^3.1.0"
   },
   "devDependencies": {
