+++ packages/examples/package.json
@@ -33,7 +33,6 @@
     "@jsonforms/webcomponent": "^2.0.0-rc.1",
     "react": "^16.2.0",
     "react-redux": "^5.0.6",
-    "redux": "^3.7.2",
-    "redux-thunk": "^2.2.0"
+    "redux": "^3.7.2"
   }
 }
