+++ packages/i18n/package.json
@@ -40,8 +40,7 @@
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
     "react-redux": "^5.0.6",
-    "redux": "^3.7.2",
-    "redux-thunk": "^2.2.0"
+    "redux": "^3.7.2"
   },
   "devDependencies": {
     "@jsonforms/test": "^2.0.0-rc.1",
