+++ packages/vanilla/package.json
@@ -31,7 +31,6 @@
     "jsdom": "^11.5.1",
     "jsdom-global": "^3.0.2",
     "react-dom": "^16.2.0",
-    "redux-thunk": "^2.2.0",
     "source-map-loader": "^0.2.3",
     "ts-loader": "^3.2.0",
     "tslint-loader": "^3.5.3",
